Create api for addNewServies and editNewServies

+++ Assistant-manager v1.0/Model/Price.swift
@@ -12,7 +12,7 @@ class Price {
     var idPrice: String!
     var nameServise: String!
     var priceServies: Double!
-    var timeAtWorkMin: Double!
+    var timeAtWorkMin: Int!
     var timeReturnServiseDays: Int!
     var ratingService: Int!
     var remoteService: Bool!
@@ -22,7 +22,7 @@ class Price {
         self.idPrice = dictionary["idPrice"] as? String ?? ""
         self.nameServise = dictionary["nameServise"] as? String ?? ""
         self.priceServies = dictionary["priceServies"] as? Double ?? nil
-        self.timeAtWorkMin = dictionary["timeAtWorkMin"] as? Double ?? nil
+        self.timeAtWorkMin = dictionary["timeAtWorkMin"] as? Int ?? nil
         self.timeReturnServiseDays = dictionary["timeReturnServiseDays"] as? Int ?? nil
         self.ratingService = dictionary["ratingService"] as? Int ?? nil
         self.remoteService = dictionary["remoteService"] as? Bool ?? nil
